+++ functional-tests/api/saga.test.ts
@@ -1,5 +1,5 @@
 import { nanoid } from 'nanoid'
-import { Client } from '@resolve-js/client'
+import { Client, createWaitForResponseMiddleware } from '@resolve-js/client'
 
 import {
   SAGA_TEST_EVENTS_COUNT,
@@ -48,11 +48,17 @@ test('executes side effects with "live" events only', async () => {
         args: {},
       },
       {
-        waitFor: {
-          validator: (result) =>
-            Array.isArray(result?.data) && result?.data.length === 1,
-          attempts: 5,
-          period: 3000,
+        middleware: {
+          response: createWaitForResponseMiddleware({
+            validator: async (response, confirm) => {
+              const result = await response.json()
+              if (Array.isArray(result?.data) && result?.data.length === 1) {
+                confirm(result)
+              }
+            },
+            attempts: 5,
+            period: 3000,
+          }),
         },
       }
     )
@@ -89,11 +95,17 @@ test('executes side effects with "live" events only', async () => {
         args: {},
       },
       {
-        waitFor: {
-          validator: (result) =>
-            Array.isArray(result?.data) && result?.data.length === 2,
-          attempts: 5,
-          period: 3000,
+        middleware: {
+          response: createWaitForResponseMiddleware({
+            validator: async (response, confirm) => {
+              const result = await response.json()
+              if (Array.isArray(result?.data) && result?.data.length === 2) {
+                confirm(result)
+              }
+            },
+            attempts: 5,
+            period: 3000,
+          }),
         },
       }
     )