+++ packages/core/core/src/aggregate/get-aggregates-interop-builder.ts
@@ -415,9 +415,7 @@ const makeCommandExecutor = (
 
     monitoringGroup?.time('Execution')
 
-    const { jwt: actualJwt, jwtToken: deprecatedJwt } = command
-
-    const jwt = actualJwt || deprecatedJwt
+    const { jwt } = command
 
     let executionError