+++ packages/runtime/adapters/eventstore-adapters/eventstore-base/src/types.ts
@@ -92,17 +92,6 @@ export function getInitialReplicationState(): ReplicationState {
   }
 }
 
-export type EventStoreDescription = {
-  eventCount: number
-  secretCount: number
-  setSecretCount: number
-  deletedSecretCount: number
-  isFrozen: boolean
-  lastEventTimestamp: number
-  cursor?: Cursor
-  resourceNames?: { [key: string]: string }
-}
-
 export type CheckForResourceError = (errors: Error[]) => void
 
 type DeleteSecret = SecretsManager['deleteSecret']
@@ -487,10 +476,6 @@ export interface AdapterFunctions<
     ConnectedProps,
     Adapter['setReplicationStatus']
   >
-  setReplicationIterator?: PoolMethod<
-    ConnectedProps,
-    Adapter['setReplicationIterator']
-  >
   setReplicationPaused?: PoolMethod<
     ConnectedProps,
     Adapter['setReplicationPaused']
@@ -522,6 +507,10 @@ export interface EventLoader {
   readonly close: () => Promise<void>
   readonly cursor: InputCursor
   readonly isNative: boolean
+}
+
+export type EventLoaderOptions = {
+  preferRegular: boolean // prefer regular implementation via loadEvents over native one
 }
 
 export interface Adapter extends CoreEventstore {
@@ -557,8 +546,10 @@ export interface Adapter extends CoreEventstore {
     untilEventTypes: Array<InputEvent['type']>
   ) => Promise<string>
 
-  describe: () => Promise<EventStoreDescription>
   establishTimeLimit: (getVacantTimeInMillis: () => number) => void
 
-  getEventLoader: (filter: EventLoaderFilter) => Promise<EventLoader>
+  getEventLoader: (
+    filter: EventLoaderFilter,
+    options?: EventLoaderOptions
+  ) => Promise<EventLoader>
 }