+++ packages/tools/testing-tools/src/flow/read-model/query.ts
@@ -8,7 +8,7 @@ import {
   TestQuery,
 } from '../../types'
 import { as } from './as'
-import { setSecretsManager, withSecretsManager } from './with-secrets-manager'
+import { withSecretsManager } from './with-secrets-manager'
 import { makeTestEnvironment } from './make-test-environment'
 import { shouldReturn } from './should-return'
 import { not } from './not'
@@ -16,7 +16,6 @@ import { not } from './not'
 export type QueryNode = {
   as: OmitFirstArgument<typeof as>
   withSecretsManager: OmitFirstArgument<typeof withSecretsManager>
-  setSecretsManager: OmitFirstArgument<typeof setSecretsManager>
   shouldReturn: OmitFirstArgument<typeof shouldReturn>
   not: OmitFirstArgument<typeof not>
 } & Promise<QueryTestResult>
@@ -42,7 +41,6 @@ export const query = (
   return Object.assign(context.environment.promise, {
     as: partial(as, context),
     withSecretsManager: partial(withSecretsManager, context),
-    setSecretsManager: partial(setSecretsManager, context),
     shouldReturn: partial(shouldReturn, context),
     not: partial(not, context),
   })